+++ conviva/src/main/java/com/bitmovin/analytics/conviva/ConvivaAnalyticsIntegration.java
@@ -5,6 +5,9 @@ import android.os.Handler;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
+import com.bitmovin.analytics.conviva.ssai.DefaultSsaiApi;
 
 import com.bitmovin.analytics.conviva.ssai.DefaultSsaiApi;
 import com.bitmovin.analytics.conviva.ssai.SsaiApi;
@@ -31,21 +34,49 @@ import java.util.Map;
 public class ConvivaAnalyticsIntegration {
     public static final String STREAM_TYPE = "streamType";
     public static final String INTEGRATION_VERSION = "integrationVersion";
-
     private static final String TAG = "ConvivaAnalyticsInt";
 
-    private final ContentMetadataBuilder contentMetadataBuilder = new ContentMetadataBuilder();
-    private final ConvivaVideoAnalytics convivaVideoAnalytics;
-    private final ConvivaAdAnalytics convivaAdAnalytics;
     @NonNull
-    private final PlayerDecorator player;
+    private final ContentMetadataBuilder contentMetadataBuilder = new ContentMetadataBuilder();
+    @NonNull
+    private final ConvivaVideoAnalytics convivaVideoAnalytics;
+    @NonNull
+    private final ConvivaAdAnalytics convivaAdAnalytics;
+
+    @Nullable
+    private PlayerDecorator player;
+
     private MetadataOverrides metadataOverrides;
+    @NonNull
     private final DefaultSsaiApi ssai;
 
-    // Helper
     private Boolean isSessionActive = false;
     private Boolean isBumper = false;
     private Boolean isBackgrounded = false;
+
+    public ConvivaAnalyticsIntegration(String customerKey, Context context) {
+        this(
+            null,
+            customerKey,
+            context,
+            new ConvivaConfig(),
+            null,
+            null,
+            null
+        );
+    }
+
+    public ConvivaAnalyticsIntegration(String customerKey, Context context, ConvivaConfig config) {
+        this(
+            null,
+            customerKey,
+            context,
+            config,
+            null,
+            null,
+            null
+        );
+    }
 
     public ConvivaAnalyticsIntegration(Player player, String customerKey, Context context) {
         this(player, customerKey, context, new ConvivaConfig());
@@ -61,15 +92,14 @@ public class ConvivaAnalyticsIntegration {
     }
 
     ConvivaAnalyticsIntegration(
-        Player player,
+        @Nullable Player player,
         String customerKey,
         Context context,
         ConvivaConfig config,
-        ConvivaVideoAnalytics videoAnalytics,
-        ConvivaAdAnalytics adAnalytics,
-        DefaultSsaiApi ssai
+        @Nullable ConvivaVideoAnalytics videoAnalytics,
+        @Nullable ConvivaAdAnalytics adAnalytics,
+        @Nullable DefaultSsaiApi ssai
     ) {
-        this.player = new DefaultPlayerDecorator(player);
         Map<String, Object> settings = new HashMap<>();
         if (config.getGatewayUrl() != null || config.isDebugLoggingEnabled()) {
             if (config.getGatewayUrl() != null) {
@@ -93,12 +123,13 @@ public class ConvivaAnalyticsIntegration {
         }
 
         if (ssai == null) {
-            this.ssai = new DefaultSsaiApi(convivaVideoAnalytics, convivaAdAnalytics, this.player);
+            this.ssai = new DefaultSsaiApi(convivaVideoAnalytics, convivaAdAnalytics);
         } else {
             this.ssai = ssai;
         }
-
-        attachBitmovinEventListeners();
+        if (player != null) {
+            attachPlayer(player);
+        }
         setUpAdAnalyticsCallback();
     }
 
@@ -106,8 +137,8 @@ public class ConvivaAnalyticsIntegration {
         convivaAdAnalytics.setCallback(new ConvivaExperienceAnalytics.ICallback() {
             @Override
             public void update() {
-                if (isAdActive()) {
-                    convivaAdAnalytics.reportAdMetric(ConvivaSdkConstants.PLAYBACK.PLAY_HEAD_TIME, ((long) (player.getPlayHeadTimeMillis())));
+                if (isAdActive() && player != null) {
+                    convivaAdAnalytics.reportAdMetric(ConvivaSdkConstants.PLAYBACK.PLAY_HEAD_TIME, player.getPlayHeadTimeMillis());
                 }
             }
 
@@ -118,7 +149,7 @@ public class ConvivaAnalyticsIntegration {
     }
 
     private boolean isAdActive() {
-        return player.isAd() || ssai.isAdBreakActive();
+        return (player != null && player.isAd()) || ssai.isAdBreakActive();
     }
 
     // region public methods
@@ -151,14 +182,15 @@ public class ConvivaAnalyticsIntegration {
      * no longer ensure that the session is managed at the correct time. Additional: Since some metadata attributes
      * relies on the players source we can't ensure that all metadata attributes are present at session creation.
      * Therefore it could be that there will be a 'ContentMetadata created late' issue after conviva validation.
-     * <p>
-     * If no source was loaded this method will throw an error.
+     * <p/>
+     * If no no player with loaded source is available, `assetName` must be set via `updateContentMetadata` before
+     * calling this method.
      */
     public void initializeSession() throws ConvivaAnalyticsException {
-        if ((player.getStreamTitle() == null)
+        if ((player == null || player.getStreamTitle() == null)
                 && this.contentMetadataBuilder.getAssetName() == null) {
             throw new ConvivaAnalyticsException(
-                    "AssetName is missing. Load player source (with Title) first or set assetName via updateContentMetadata"
+                    "AssetName is missing. Load player source (with Title) and attach player first or set assetName via updateContentMetadata"
             );
         }
         internalInitializeSession();
@@ -200,7 +232,9 @@ public class ConvivaAnalyticsIntegration {
     public void release(Boolean releaseConvivaSdk) {
         convivaAdAnalytics.release();
         convivaVideoAnalytics.release();
-        detachBitmovinEventListeners();
+        if (player != null) {
+            detachBitmovinEventListeners(player);
+        }
         if (releaseConvivaSdk) {
             ConvivaAnalytics.release();
         }
@@ -278,6 +312,33 @@ public class ConvivaAnalyticsIntegration {
 
     }
 
+    /**
+     * Attaches a `Player` instance to the Conviva Analytics object.
+     * This method should be called as soon as the `Player` instance is initialized to not miss any tracking.
+     *
+     * Has no effect if there is already a `Player` instance set. Use the `ConvivaAnalyticsIntegration` constructor
+     * without `player` if you plan to attach a `Player` instance later in the life-cycle.
+     */
+    public void attachPlayer(@NonNull Player player) {
+        if (this.player != null) {
+            Log.w(TAG, "There is already a Player instance attached! Ignoring new Player instance.");
+            return;
+        }
+
+        if (player.getSource() != null) {
+            Log.w(
+                TAG,
+                "Player already has a source loaded. Please provide the player instance before loading a source."
+            );
+        }
+
+        this.player = new DefaultPlayerDecorator(player);
+        updateSession();
+
+        attachBitmovinEventListeners(this.player);
+        ssai.setPlayer(this.player);
+    }
+
     // endregion
 
     private void ensureConvivaSessionIsCreatedAndInitialized() {
@@ -306,11 +367,11 @@ public class ConvivaAnalyticsIntegration {
     }
 
     private void internalInitializeSession() {
-        if (isSessionActive) {
-            return;
-        }
+        if (isSessionActive) return;
+
         Log.d(TAG, "internalInitializeSession");
         createContentMetadata();
+
         convivaVideoAnalytics.reportPlaybackRequested(contentMetadataBuilder.build());
         setupPlayerStateManager();
         if (metadataOverrides != null) {
@@ -320,15 +381,17 @@ public class ConvivaAnalyticsIntegration {
     }
 
     private void updateSession() {
-        updatePlaybackVideoData();
-        buildDynamicContentMetadata();
+        if (player != null) {
+            updatePlaybackVideoData(player);
+            buildDynamicContentMetadata(player);
+        }
 
         if (isSessionActive) {
             convivaVideoAnalytics.setContentInfo(contentMetadataBuilder.build());
         }
     }
 
-    private void updatePlaybackVideoData() {
+    private void updatePlaybackVideoData(@NonNull PlayerDecorator player) {
         HashMap<String, Object[]> playbackVideoData = player.getPlaybackVideoData();
         for (Map.Entry<String, Object[]> entry : playbackVideoData.entrySet()) {
             convivaVideoAnalytics.reportPlaybackMetric(entry.getKey(), entry.getValue());
@@ -343,17 +406,19 @@ public class ConvivaAnalyticsIntegration {
         if (overriddenAssetName != null) {
             contentMetadataBuilder.setAssetName(overriddenAssetName);
         } else {
-            if (player.getStreamTitle() != null) {
+            if (player != null && player.getStreamTitle() != null) {
                 contentMetadataBuilder.setAssetName(player.getStreamTitle());
             } else {
                 Log.w(TAG, "No asset name provided for content metadata.");
             }
         }
 
-        buildDynamicContentMetadata();
+        if (player != null) {
+            buildDynamicContentMetadata(player);
+        }
     }
 
-    private void buildDynamicContentMetadata() {
+    private void buildDynamicContentMetadata(PlayerDecorator player) {
         // Build custom tags here, though this is static metadata but
         // streamType could be missing at time of session initialization
         // as source information could be unavailable at that time
@@ -384,7 +449,7 @@ public class ConvivaAnalyticsIntegration {
     }
     // endregion
 
-    private void attachBitmovinEventListeners() {
+    private void attachBitmovinEventListeners(PlayerDecorator player) {
         player.withEventEmitter(eventEmitter -> {
             eventEmitter.on(SourceEvent.Unloaded.class, onSourceUnloadedListener);
             eventEmitter.on(PlayerEvent.Error.class, onPlayerErrorListener);
@@ -424,7 +489,7 @@ public class ConvivaAnalyticsIntegration {
         });
     }
 
-    private void detachBitmovinEventListeners() {
+    private void detachBitmovinEventListeners(PlayerDecorator player) {
         player.withEventEmitter(bitmovinPlayer -> {
             bitmovinPlayer.off(SourceEvent.Unloaded.class, onSourceUnloadedListener);
             bitmovinPlayer.off(PlayerEvent.Error.class, onPlayerErrorListener);
@@ -493,12 +558,9 @@ public class ConvivaAnalyticsIntegration {
         }
     };
 
-    private final EventListener<SourceEvent.Error> onSourceErrorListener = new EventListener<SourceEvent.Error>() {
-        @Override
-        public void onEvent(SourceEvent.Error event) {
-            Log.d(TAG, "[Source Event] Error");
-            handleError(String.format("%s - %s", event.getCode(), event.getMessage()));
-        }
+    private final EventListener<SourceEvent.Error> onSourceErrorListener = event -> {
+        Log.d(TAG, "[Source Event] Error");
+        handleError(String.format("%s - %s", event.getCode(), event.getMessage()));
     };
 
     private void handleError(String message) {
@@ -544,7 +606,6 @@ public class ConvivaAnalyticsIntegration {
         ensureConvivaSessionIsCreatedAndInitialized();
         updateSession();
     };
-
     private final EventListener<PlayerEvent.Playing> onPlayingListener = playingEvent -> {
         Log.d(TAG, "[Player Event] Playing");
         contentMetadataBuilder.setPlaybackStarted(true);
@@ -631,7 +692,7 @@ public class ConvivaAnalyticsIntegration {
         // Notify of seek buffering complete at this stage.
         Log.d(TAG, "[Player Event] Update state after buffering");
         ConvivaSdkConstants.PlayerState state = ConvivaSdkConstants.PlayerState.PAUSED;
-        if (player.isPlaying()) {
+        if (player != null && player.isPlaying()) {
             state = ConvivaSdkConstants.PlayerState.PLAYING;
         }
         transitionState(state);
@@ -755,7 +816,7 @@ public class ConvivaAnalyticsIntegration {
         ConvivaSdkConstants.AdPosition adPosition = ConvivaSdkConstants.AdPosition.MIDROLL;
         if (timeOffset == 0.0) {
             adPosition = ConvivaSdkConstants.AdPosition.PREROLL;
-        } else if (timeOffset == player.getDuration()) {
+        } else if (player != null && timeOffset == player.getDuration()) {
             adPosition = ConvivaSdkConstants.AdPosition.POSTROLL;
         }
         return adPosition;
@@ -800,5 +861,4 @@ public class ConvivaAnalyticsIntegration {
             }
         }
     };
-    // endregion
 }